+++ euporie/core/comm/__init__.py
@@ -1,1 +1,3 @@
 """Sub-module for handling kernel Comm messages."""
+
+from . import ipympl as ipympl